Extract Ruby declarator partial
